# backend/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Amana Backend API",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Current versioned lane (recommended)"
    },
    {
      "url": "http://localhost:4000",
      "description": "Deprecated legacy alias (serves Deprecation/Sunset headers)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-secret",
        "description": "Secret key authentication for admin-only management routes."
      }
    },
...
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "example": {
          "error": "Unauthorized"
        }
      },
      "AppErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "AUTH_ERROR",
              "DOMAIN_ERROR",
              "INFRA_ERROR",
              "NOT_FOUND",
              "INTERNAL_ERROR",
              "CLAWBACK_TOO_LARGE",
              "CLAWBACK_INVALID_AMOUNT",
              "SUBMISSION_VALIDATION_ERROR",
              "SUBMISSION_NETWORK_ERROR",
              "SUBMISSION_CONTRACT_ERROR",
              "SUBMISSION_AUTHORIZATION_ERROR"
            ],
            "description": "Error code classifying the failure type. Submission error codes:\n- `SUBMISSION_VALIDATION_ERROR`: Malformed transaction XDR (400)\n- `SUBMISSION_NETWORK_ERROR`: Network timeout, connection refused, or rate limiting (429/502/503/504)\n- `SUBMISSION_CONTRACT_ERROR`: Smart contract rejected the transaction (404/502)\n- `SUBMISSION_AUTHORIZATION_ERROR`: Transaction was not authorized (401/403)"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "oneOf": [
              {
                "type": "object"
              },
              {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            ]
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "example": {
          "code": "VALIDATION_ERROR",
          "message": "Validation failed",
          "details": [
            {
              "path": "body.amountUsdc",
              "message": "Invalid amount format"
            }
          ]
        }
      },
...
      "UnsignedXdrResponse": {
        "type": "object",
        "properties": {
          "unsignedXdr": {
            "type": "string"
          }
        },
        "required": [
          "unsignedXdr"
        ]
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Missing, invalid, revoked, or semantically invalid bearer token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "unauthorized": {
                "value": {
                  "error": "Unauthorized"
                }
              },
              "revoked": {
                "value": {
                  "error": "Unauthorized: token has been revoked"
                }
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Authenticated caller is not allowed to perform the action.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": "Forbidden"
            }
          }
        }
      },
...
      "LegacyValidationError": {
        "description": "Request payload failed route-level validation.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              ]
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected backend failure.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              ]
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/admin/features": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List system feature flags",
        "description": "Get active feature flags and system runtime toggles for administrative controls.",
        "security": [
          {
            "adminAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Feature flag status map returned."
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
    "/api/admin/contract/mediators": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Register a mediator on the escrow contract",
        "description": "Builds an unsigned Soroban transaction that adds `mediatorAddress` to the\non-chain mediator registry (`add_mediator`), previously only doable via the\n`soroban contract invoke` CLI. Admin-only; the caller's wallet is the\ntransaction source/signer. As with other admin transaction endpoints, the\nbackend never holds a signing key — the caller signs and submits the\nreturned XDR themselves.\nError code references: `AUTH_ERROR`, `VALIDATION_ERROR`, `INTERNAL_ERROR`.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mediatorAddress": {
                    "type": "string",
                    "description": "Stellar public key to register as a mediator."
                  }
                },
                "required": [
                  "mediatorAddress"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Add-mediator transaction built.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnsignedXdrResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/LegacyValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/admin/contract/mediators/{address}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Revoke a mediator on the escrow contract",
        "description": "Builds an unsigned Soroban transaction that removes `address` from the\non-chain mediator registry (`remove_mediator`). Admin-only; the caller's\nwallet is the transaction source/signer.\nError code references: `AUTH_ERROR`, `VALIDATION_ERROR`, `INTERNAL_ERROR`.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "description": "Stellar public key to remove from the mediator registry.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Remove-mediator transaction built.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnsignedXdrResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/LegacyValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/admin/contract/fee": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Update the escrow contract platform fee rate",
        "description": "Builds an unsigned Soroban transaction that updates the platform fee\n(`update_fee_bps`). The contract itself enforces the fee stays within its\nconfigured [MIN_FEE_BPS, MAX_FEE_BPS] bounds; this endpoint additionally\nrejects out-of-range values before building a transaction. Admin-only;\nthe caller's wallet is the transaction source/signer.\nError code references: `AUTH_ERROR`, `VALIDATION_ERROR`, `INTERNAL_ERROR`.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "feeBps": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 500,
                    "description": "New platform fee rate in basis points."
                  }
                },
                "required": [
                  "feeBps"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fee-update transaction built.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnsignedXdrResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/LegacyValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  }
}